[BACKPORT 2024.1][PLAT-13651]LDAP sync should have an option to extract users from other attributes as well
Summary:
Original commit: e80bcd00aecd56d89edbbd03082d4b3d5e3dba43 / D34518
Currently during our YBA ↔︎ YBDB LDAP sync, we assume that the user name we want to sync will be present on the DN. While that will be true for most of the scenarios, but we can have customers wanting to sync the user present on a different attribute for example - `sAMAccountName`

This diff performs the sync based on the attribute the user specified in the payload. If the specified `ldapUserfield` is not present in the `DN`, the user name will be retrieved from this attribute on the LDAP server. If the attribute is also not found, the user is simply skipped from the sync.

Test Plan:
Manual testing
  - Triggered the sync with the ldapUserfield and observed the sync where the user name is retrieved from the dn
  - Triggered the sync with the ldapUserfield [not present on the DN] and synced only the users that have this attribute set on the LDAP server

Reviewers: #yba-api-review!, svarshney

Reviewed By: svarshney

Subscribers: yugaware

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D34625