[BACKPORT 2.20][PLAT-13681] Generate YBA Managed bundles in case the provider contains legacy bundles
Summary:
Before the VM_OS_Patching flag is enabled we do not set the versions in the `yugaware_property` because of which the
bundles were not getting generated even in case YBA moved to a higher version.

With this diff, we fix the same & generates the bundle.
1. If the version specified in the `yugaware_property` does not match the one present in `CloudImageBundleSetup`.
2. Bundles are legacy one without the metadata.

Depends on - D34469
Original Commit - https://github.com/yugabyte/yugabyte-db/commit/c4056b7ad8065d2f057d555be01c1916c8d7bf3f

Test Plan:
Created a provider in YBA 2.18
Migrated the same to the YBA version having the patch applied.
Verified that the bundles are getting generated as expected.

Changed the version in YBA metadata manually
Verified that the bundles are generated as expected.

Reviewers: amalyshev

Reviewed By: amalyshev

Subscribers: yugaware

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D34581