[BACKPORT 2024.1][#22211] xCluster: Create new DB Scoped stream in INITIATED state
Summary:
Original commit: 59790d3935c2d6d0f9eebfd859e5597ea6174e25 / D34650
Streams should be in INITIATED state until replication starts. This is done in order to block tablet splitting while we are in the setup phase.
The backup restore which sets up the tablets on the target cluster. When xCluster setup expects the tablet count to be the same to simplify the process.
Jira: DB-11131

Test Plan: Jenkins

Reviewers: jhe, xCluster

Reviewed By: jhe

Subscribers: ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D34654