[BACKPORT 2024.1][PLAT-13718] Preselect tables in error state for restart replication
Summary:
**Context**
Restart replication (also known as 'resync DR replica') is an operation for
xCluster configurations (both xCluster replication and xCluster DR) which is
performed when a backup of the source (DR primary) and restore on the target (DR
replica) is desired.
On the restart replication modal, the user would select all tables in `ERROR`
state in most cases. The `error` state indicates that the write-ahead logs for
these tables are already deleted so replication restart (backup & restore) is
required for replication to continue on these tables.

Thus, it makes sense for the YBA UI to preselect these tables for the user to
make this a smoother experience.
For YSQL xCluster replications, we can only do backup and restore at the database
level. This means we will need to preselect any database which contains tables in
`error` state.

Original commit: 5c66edcddbe91b3efe1f5166031d2c53eca4927c / D34838

**Changes**
- Added logic to preselect databases (for YSQL) / tables (for YCQL) based on the
table `error` status.

Test Plan:
- Mark a table with error status and ensure that only the databases which
contain tables in error status are selected and submitted to the backend as part
of the restart replication request.
- If no tables are in error status, then no tables are preselected.

{F176434}

Reviewers: rmadhavan, kkannan, hzare, cwang

Reviewed By: cwang

Subscribers: yugaware

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35058