[BACKPORT 2024.1][#22423] DocDB: Clarify that memory division flags are a percentage of hard memory limit
Summary:
Original commit: 4801c923ce56756934af6666f174aae63b791fa0 / D35135

clean cherry pick

Some of our memory division gflags just say "percentage of total available memory" or the like, which is misleading. They are actually a percentage of the process is hard memory limit, which in turn is a percentage of the total available memory on the box.

This fixes that.
Jira: DB-11333

Test Plan: eyeballs

Reviewers: asrivastava

Reviewed By: asrivastava

Subscribers: ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D35141